+++ styles.py
@@ -1,21 +1,25 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
 # this file contains the main styling of the application (e.g. butttons, input field etc.)
 
-# global width and height
-globalWidth = 762 # fixed length
-globalHeight = 275 # fixed length
+# global width and height of application
+globalWidth = 562 # fixed length
+globalHeight = 600 # fixed length
+
+# background colour of application
 backgroundcolour = '#A9E2F3'
 
-# width of note for responsiveness
-widgetWidth = 650 # it seems that this is obsolete --> test!
-#widgetWidth = globalWidth - 110
-#widgthHeight = globalHeight - 50
+# application logo
+app_logo = './Logo/applogo.png'
+
 
 
 # general label styling
 label = {
-    'x-axis': 0,
+    'x-axis': 1,
     'width': 200,
-    'height': 25
+    'height': 25,
 }
 
 
@@ -26,8 +30,18 @@ input = {
     'height': 25
 }
 
+
 # general button styling
 button = {
     'width': 200,
     'height': 25
+}
+
+
+# text field for query information
+text = {
+    'x-axis': 0,
+    'y-axis': 310,
+    'width': 629,
+    'height': 274,
 }